It looks finished!

+++ CorporateCustomer.java
@@ -1,6 +1,6 @@
 public class CorporateCustomer extends Customer{
     private String contactName;
-    public CorporateCustomer(String contact, String customerName, String caddress, float newcred, float newdiscount){
+    public CorporateCustomer(String contact, String customerName, String caddress, float newcred, double newdiscount){
         contactName = contact;
         name = customerName;
         address = caddress;
